Refactor code to have most of the logic in AbstractGitRepository
Now only the method `getCheckoutUri` needs to be implemented for each git repository class
