+++ src/main/scala/io/viash/functionality/dependencies/GitRepository.scala
@@ -63,4 +63,7 @@ case class GitRepository(
   ): GitRepository = {
     copy(name, `type`, uri, tag, path, localPath)
   }
+
+  def getCheckoutUri(): String = uri
+
 }
